python-requirements.txt
wheel
Django
BeautifulSoup4
requests
mysqlclient
networkx
django-background-tasks
